// frontend/src/api/model/serverItemStatus.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * DevOps Dashboard API
 * 运维监控仪表盘一期 API 规范（Spec-Driven Development）
 * OpenAPI spec version: 1.0.0
 */

export type ServerItemStatus = typeof ServerItemStatus[keyof typeof ServerItemStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ServerItemStatus = {
  running: 'running',
  stopped: 'stopped',
  maintenance: 'maintenance',
} as const;